pyproject.toml
[project]
name = "Quart-SocketIO"
version = "0.1.0"
authors = [
    { name = "Nicholas Silva", email = "nicholas@robotz.dev" },
    { name = "Miguel Grinberg", email = "miguel.grinberg@gmail.com" },
]
description = "Async Socket.IO integration for Quart applications"
classifiers = [
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.13"
dependencies = [
    "quart",
    "python-socketio >= 5.12.0",
    "uvicorn (>=0.34.3)",
    "hypercorn>=0.18.0",
]


[project.urls]
Homepage = "https://github.com/Robotz213/Quart-SocketIO"
"Bug Tracker" = "https://github.com/Robotz213/Quart-SocketIO/issues"

[project.readme]
file = "README.md"
content-type = "text/markdown"


[tool.setuptools]
zip-safe = false
include-package-data = true


[build-system]
requires = ["setuptools>=61.2"]
build-backend = "setuptools.build_meta"
